fix: move GetMockTL to tlog/mock.go
